requirements.txt
kaggle
python-dotenv
requests
